Merge feature slash commands from origin/experimental (Alpha-0.2.6) (#6)
* Update README.md

* Change notices for experimental version

* Move all  listeners off the project
As of build e5, the traditional listeners of all commands will be have to be moved off the repository to add slash commands. The easter egg commands will be dealt with separately
I have already started work on the info command

* Add build info
This commit adds build info to the experimental branch
It will be printed at the start of the main class.

This commit does contain the new listeners, but they are not working in this commit, but should be sorted out in the next commit.

Lavaplayer has been disabled (as a dependency) because it is not be implemented on this branch yet

* Cosmetic Updates

This adds version variables to the core and changes console logger colors. Also invalidates the need for a networking class at this time

* Build Updates

Update the JDA branch to slash-commands because buttons is not a branch anymore

* Add meta

Allow build information to be accessed via variables

* Cosmetics Update 2

This restyles the logger. Continuation of 0511a693

* Cosmetics Update 2

This restyles the logger. Continuation of 0511a693

* Complete Removal Of Old Listeners

* Delete runoff of merge conflict fix

* WIDGET LOL

* WIDGET DEAD LOL I GIVE UP

* Add Moderation Features

This adds the ban, unban, and kick commands to the bot, as well as the info command

This is the 2nd-to-last step before merging slash commands into stable

* Final Preparations For Merging